fix: add WindIndicator style to App.axaml
WindIndicator style was moved from WeatherPlugin.cs to App.axaml. This change was necessary because we are refactoring styles and resources to allow for easier style management and cohesion in a centralized location. Loading the style at the application level ensures the style is available to all components, reducing redundant code.

Asana: https://app.asana.com/0/1203851531040615/1204811123850722/f