Add error handling and localization for weather providers
This commit introduces error messages and localizes them for the WindyProvider and OpenWeatherMapProvider. When the HTTP request to fetch weather data fails or throws an exception, an error message will be logged and displayed to the user. This helps in debugging and also improves the user experience as users will be informed about the issues in their language. The error messages have been added to the Russian and default resource files. In addition, the base initialization in WeatherActionViewModel has been included for proper setup of the object.

Asana: https://app.asana.com/0/1203851531040615/1205134810486883/f